Remove travis.yml from repo
There are no tests in the repo, so we have no need for CI. I know what
you're probably thinking...why not add some tests, right? I hear ya.
That thought is at the back of my mind as well. I'm asking the same
thing about all previous commits, too, but for the time being, the day
is short and the backlog long, so this can gets kicked down the road yet
again. #forshame